requirements.txt
fastembed
opencv-python
pdfplumber
python-dotenv
PyMuPDF
pytesseract
torch
transformers
qdrant_client
llama-index
llama-index-embeddings-huggingface
llama-index-llms-openai
llama-index-llms-together
llama-index-llms-anthropic
llama-index-vector-stores-qdrant